+++ src/main/java/com/kalsym/order/service/model/Order.java
@@ -129,6 +129,8 @@ public class Order implements Serializable {
     
     private String orderGroupId;
     
+    private String orderQrGroupId;
+    
     @Enumerated(EnumType.STRING)
     private ServiceType serviceType;