+++ src/main/java/com/kalsym/order/service/model/OrderGroup.java
@@ -134,5 +134,5 @@ public class OrderGroup implements Serializable {
             return null;
         }
     }
-   
+        
 }