+++ src/main/java/com/kalsym/order/service/model/StoreWithDetails.java
@@ -204,4 +204,6 @@ public class StoreWithDetails implements Serializable {
         }
         return null;
     }
+    
+    private Boolean dineInConsolidatedOrder;
 }